% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_functions.R
\name{pagespeedapi.runpagespeed}
\alias{pagespeedapi.runpagespeed}
\title{Runs PageSpeed analysis on the page at the specified URL, and returns PageSpeed scores, a list of suggestions to make that page faster, and other information.}
\usage{
pagespeedapi.runpagespeed(url, filter_third_party_resources = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, strategy = NULL)
}
\arguments{
\item{url}{The URL to fetch and analyze}

\item{filter_third_party_resources}{Indicates if third party resources should be filtered out before PageSpeed analysis}

\item{locale}{The locale used to localize formatted results}

\item{rule}{A PageSpeed rule to run; if none are given, all rules are run}

\item{screenshot}{Indicates if binary data containing a screenshot should be included}

\item{strategy}{The analysis strategy to use}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/speed/docs/insights/v2/getting-started}{Google Documentation}
}
