% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{PagespeedApiFormatStringV2.args}
\alias{PagespeedApiFormatStringV2.args}
\title{PagespeedApiFormatStringV2.args Object}
\usage{
PagespeedApiFormatStringV2.args(PagespeedApiFormatStringV2.args.rects = NULL,
  PagespeedApiFormatStringV2.args.secondary_rects = NULL)
}
\arguments{
\item{PagespeedApiFormatStringV2.args.rects}{The \link{PagespeedApiFormatStringV2.args.rects} object or list of objects}

\item{PagespeedApiFormatStringV2.args.secondary_rects}{The \link{PagespeedApiFormatStringV2.args.secondary_rects} object or list of objects}
}
\value{
PagespeedApiFormatStringV2.args object
}
\description{
PagespeedApiFormatStringV2.args Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PagespeedApiFormatStringV2 functions: \code{\link{PagespeedApiFormatStringV2.args.rects}},
  \code{\link{PagespeedApiFormatStringV2.args.secondary_rects}},
  \code{\link{PagespeedApiFormatStringV2}}
}
