% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{PagespeedApiFormatStringV2.args.rects}
\alias{PagespeedApiFormatStringV2.args.rects}
\title{PagespeedApiFormatStringV2.args.rects Object}
\usage{
PagespeedApiFormatStringV2.args.rects()
}
\value{
PagespeedApiFormatStringV2.args.rects object
}
\description{
PagespeedApiFormatStringV2.args.rects Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PagespeedApiFormatStringV2 functions: \code{\link{PagespeedApiFormatStringV2.args.secondary_rects}},
  \code{\link{PagespeedApiFormatStringV2.args}},
  \code{\link{PagespeedApiFormatStringV2}}
}
