% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{PagespeedApiFormatStringV2}
\alias{PagespeedApiFormatStringV2}
\title{PageSpeed Insights API Objects 
Analyzes the performance of a web page and provides tailored suggestions to make that page faster.}
\usage{
PagespeedApiFormatStringV2(PagespeedApiFormatStringV2.args = NULL,
  PagespeedApiFormatStringV2.args.rects = NULL,
  PagespeedApiFormatStringV2.args.secondary_rects = NULL, args = NULL,
  format = NULL)
}
\arguments{
\item{PagespeedApiFormatStringV2.args}{The \link{PagespeedApiFormatStringV2.args} object or list of objects}

\item{PagespeedApiFormatStringV2.args.rects}{The \link{PagespeedApiFormatStringV2.args.rects} object or list of objects}

\item{PagespeedApiFormatStringV2.args.secondary_rects}{The \link{PagespeedApiFormatStringV2.args.secondary_rects} object or list of objects}

\item{args}{List of arguments for the format string}

\item{format}{A localized format string with {{FOO}} placeholders, where 'FOO' is the key of the argument whose value should be substituted}
}
\value{
PagespeedApiFormatStringV2 object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:24:56
filename: /Users/mark/dev/R/autoGoogleAPI/googlepagespeedonlinev2.auto/R/pagespeedonline_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
PagespeedApiFormatStringV2 Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PagespeedApiFormatStringV2 functions: \code{\link{PagespeedApiFormatStringV2.args.rects}},
  \code{\link{PagespeedApiFormatStringV2.args.secondary_rects}},
  \code{\link{PagespeedApiFormatStringV2.args}}
}

