% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{PagespeedApiImageV2.page_rect}
\alias{PagespeedApiImageV2.page_rect}
\title{PagespeedApiImageV2.page_rect Object}
\usage{
PagespeedApiImageV2.page_rect(height = NULL, left = NULL, top = NULL,
  width = NULL)
}
\arguments{
\item{height}{The height of the rect}

\item{left}{The left coordinate of the rect, in page coordinates}

\item{top}{The top coordinate of the rect, in page coordinates}

\item{width}{The width of the rect}
}
\value{
PagespeedApiImageV2.page_rect object
}
\description{
PagespeedApiImageV2.page_rect Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The region of the page that is captured by this image, with dimensions measured in CSS pixels.
}
\seealso{
Other PagespeedApiImageV2 functions: \code{\link{PagespeedApiImageV2}}
}
