% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{PagespeedApiImageV2}
\alias{PagespeedApiImageV2}
\title{PagespeedApiImageV2 Object}
\usage{
PagespeedApiImageV2(PagespeedApiImageV2.page_rect = NULL, data = NULL,
  height = NULL, key = NULL, mime_type = NULL, page_rect = NULL,
  width = NULL)
}
\arguments{
\item{PagespeedApiImageV2.page_rect}{The \link{PagespeedApiImageV2.page_rect} object or list of objects}

\item{data}{Image data base64 encoded}

\item{height}{Height of screenshot in pixels}

\item{key}{Unique string key, if any, identifying this image}

\item{mime_type}{Mime type of image data (e}

\item{page_rect}{The region of the page that is captured by this image, with dimensions measured in CSS pixels}

\item{width}{Width of screenshot in pixels}
}
\value{
PagespeedApiImageV2 object
}
\description{
PagespeedApiImageV2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PagespeedApiImageV2 functions: \code{\link{PagespeedApiImageV2.page_rect}}
}
