% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_functions.R
\docType{package}
\name{pagespeedonline_googleAuthR}
\alias{pagespeedonline_googleAuthR}
\alias{pagespeedonline_googleAuthR-package}
\title{PageSpeed Insights API
Analyzes the performance of a web page and provides tailored suggestions to make that page faster.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:24:56
filename: /Users/mark/dev/R/autoGoogleAPI/googlepagespeedonlinev2.auto/R/pagespeedonline_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

