% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.formattedResults}
\alias{Result.formattedResults}
\title{Result.formattedResults Object}
\usage{
Result.formattedResults(Result.formattedResults.ruleResults = NULL,
  locale = NULL, ruleResults = NULL)
}
\arguments{
\item{Result.formattedResults.ruleResults}{The \link{Result.formattedResults.ruleResults} object or list of objects}

\item{locale}{The locale of the formattedResults, e}

\item{ruleResults}{Dictionary of formatted rule results, with one entry for each PageSpeed rule instantiated and run by the server}
}
\value{
Result.formattedResults object
}
\description{
Result.formattedResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Localized PageSpeed results. Contains a ruleResults entry for each PageSpeed rule instantiated and run by the server.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults.ruleResults}},
  \code{\link{Result.pageStats}},
  \code{\link{Result.ruleGroups}},
  \code{\link{Result.version}}, \code{\link{Result}}
}
