% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.pageStats}
\alias{Result.pageStats}
\title{Result.pageStats Object}
\usage{
Result.pageStats(cssResponseBytes = NULL, flashResponseBytes = NULL,
  htmlResponseBytes = NULL, imageResponseBytes = NULL,
  javascriptResponseBytes = NULL, numberCssResources = NULL,
  numberHosts = NULL, numberJsResources = NULL, numberResources = NULL,
  numberStaticResources = NULL, otherResponseBytes = NULL,
  textResponseBytes = NULL, totalRequestBytes = NULL)
}
\arguments{
\item{cssResponseBytes}{Number of uncompressed response bytes for CSS resources on the page}

\item{flashResponseBytes}{Number of response bytes for flash resources on the page}

\item{htmlResponseBytes}{Number of uncompressed response bytes for the main HTML document and all iframes on the page}

\item{imageResponseBytes}{Number of response bytes for image resources on the page}

\item{javascriptResponseBytes}{Number of uncompressed response bytes for JS resources on the page}

\item{numberCssResources}{Number of CSS resources referenced by the page}

\item{numberHosts}{Number of unique hosts referenced by the page}

\item{numberJsResources}{Number of JavaScript resources referenced by the page}

\item{numberResources}{Number of HTTP resources loaded by the page}

\item{numberStaticResources}{Number of static (i}

\item{otherResponseBytes}{Number of response bytes for other resources on the page}

\item{textResponseBytes}{Number of uncompressed response bytes for text resources not covered by other statistics (i}

\item{totalRequestBytes}{Total size of all request bytes sent by the page}
}
\value{
Result.pageStats object
}
\description{
Result.pageStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Summary statistics for the page, such as number of JavaScript bytes, number of HTML bytes, etc.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults.ruleResults}},
  \code{\link{Result.formattedResults}},
  \code{\link{Result.ruleGroups}},
  \code{\link{Result.version}}, \code{\link{Result}}
}
