% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.ruleGroups}
\alias{Result.ruleGroups}
\title{Result.ruleGroups Object}
\usage{
Result.ruleGroups()
}
\value{
Result.ruleGroups object
}
\description{
Result.ruleGroups Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map with one entry for each rule group in these results.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults.ruleResults}},
  \code{\link{Result.formattedResults}},
  \code{\link{Result.pageStats}},
  \code{\link{Result.version}}, \code{\link{Result}}
}

