% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.version}
\alias{Result.version}
\title{Result.version Object}
\usage{
Result.version(major = NULL, minor = NULL)
}
\arguments{
\item{major}{The major version number of PageSpeed used to generate these results}

\item{minor}{The minor version number of PageSpeed used to generate these results}
}
\value{
Result.version object
}
\description{
Result.version Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The version of PageSpeed used to generate these results.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults.ruleResults}},
  \code{\link{Result.formattedResults}},
  \code{\link{Result.pageStats}},
  \code{\link{Result.ruleGroups}}, \code{\link{Result}}
}

